requirements.txt
# Core dependencies
pydantic>=2.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
tenacity>=8.2.0
loguru>=0.7.0

# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# CLI
typer>=0.9.0
rich>=13.7.0

# Storage
cachetools>=5.3.0
redis>=5.0.0

# Optional: For ENS resolution and address validation
web3>=6.0.0
eth-utils>=2.3.0

# Solana dependencies
base58>=2.1.0

# Browser automation (for scraping/fallbacks)
selenium>=4.0.0

